requirements.txt
streamlit==1.31.*
chromadb==0.4.22
langchain==0.1.0
langchain-community>=0.0.20
sentence-transformers==2.3.1
openai>=1.12,<2
PyPDF2==3.0.1
python-docx==1.1.0
jieba==0.42.1
pydantic>=2.6,<2.8
python-dotenv>=1.0,<2
numpy>=1.26,<2
